config: move type out of test directory
The custom_aggregate type is used only in testing, but it seemingly
needs to be available to property<T> outside of tests in order to use
the type as a configuration option. So we define it outside of the tests
even though it is only used in testing.
