+++ src/v/config/rjson_serialization.h
@@ -15,7 +15,6 @@
 #include "config/data_directory_path.h"
 #include "config/endpoint_tls_config.h"
 #include "config/seed_server.h"
-#include "config/tests/custom_aggregate.h"
 #include "config/tls_config.h"
 #include "json/json.h"
 #include "json/stringbuffer.h"
@@ -23,6 +22,24 @@
 #include "pandaproxy/schema_registry/schema_id_validation.h"
 
 #include <seastar/core/sstring.hh>
+
+/*
+ * this type is only used for config/tests but the rjson_serialize overload
+ * needs to be here so it can be picked up by the configuration/property bits
+ * (at least without some additional c++ tricks).
+ */
+namespace testing {
+struct custom_aggregate {
+    ss::sstring string_value;
+    int int_value;
+
+    bool operator==(const custom_aggregate& rhs) const {
+        return string_value == rhs.string_value && int_value == rhs.int_value;
+    }
+
+    static consteval std::string_view type_name() { return "custom_aggregate"; }
+};
+} // namespace testing
 
 namespace json {
 
@@ -43,7 +60,7 @@ void rjson_serialize(
   const std::vector<config::seed_server>& v);
 
 void rjson_serialize(
-  json::Writer<json::StringBuffer>& w, const custom_aggregate& v);
+  json::Writer<json::StringBuffer>& w, const testing::custom_aggregate& v);
 
 void rjson_serialize(
   json::Writer<json::StringBuffer>& w, const config::endpoint_tls_config& v);
